requirements.txt
pytz
requests
python-dotenv
click==7
ipython

# Django
django==3.2
django-cleanup
django-timezone-field==4.2.1
django-cors-headers
pillow==8.2.0
django-debug-toolbar
whitenoise
django-import-export
openpyxl

# Django 3.2 async requirements
gunicorn==20.1.0
uvicorn
uvloop
httptools
django-extensions

# Databases
psycopg2-binary
dj-database-url

# Distributed tasks
celery==5.0.5
redis
django-celery-beat

# Telegram
python-telegram-bot

# monitoring
# sentry-sdk
